FIX: find existing category when a repository is moved/renamed (#52)
When a repository is moved or renamed the plugin creates a new category
on subsequent commit because the category is identified by repository
full_name (`{github_org}/{github_repo}`). This commit links the category
to repository ID so that when the repository full_name is changed it can
find the category via repository ID. It also updates the repository
full_name to reflect new location/name.