DEV: Add compatibility for Rails 7.0+
As we’re upgrading Discourse to Rails 7.0.1, the way the autoloader was
working previously changed a bit. We have weird stuff in some places
like requiring files from `lib/` with the `lib` prefix which shouldn’t
happen.

As the current version of Discourse already autoload its `lib` folder,
the `require_dependency` calls we have in this plugin aren’t useful
anymore.

This patch removes these calls which leaves the plugin compatible with
the current version of Discourse and makes it compatible with the
changes that are going to be introduced when upgrading to Rails 7.0.1.