DEV: Log debug info about auto GitHub approvals (#128)
After an approved PR is merged the commit topic that is created for the
associated commit is sometimes not automatically approved. We originally
thought it was a race condition because of reused Octokit client
instances, but that was not the problem. The additional debug
information will give more insight about what is happening inside the
auto approval process.