FIX: Escape Git trailers present in last paragraph (#131)
The previous regular expression looked for Git trailers in the whole
commit message, not only in the last paragraph. For example, it escaped
the last paragraph if the first paragraph was "foo: bar".

The new detection method removes the multiline modifier from the
regular expression and explicitly checks just the last paragraph.