FIX: Wait 30 seconds before syncing commits
Some commits are not automatically approved even if they follow all the
requirements necessary for that. Running the same sync process again at
a later time seems to work as expected.

This commits adds a 30 seconds delay between the time the merge webhook
is received and the time the sync process happens. If this is just a
timing issue, this delay should fix that.