FIX: Rename `can_review` to `can_review_code` (#141)
`CurrentUserSerializer` in core has a `can_review` attribute that's used to check if the user can see the Review link in the hamburger menu:

https://github.com/discourse/discourse/blob/010bb20f53cc81068824fc278c457ef1739f25e7/app/serializers/current_user_serializer.rb#L263-L265

However, this plugin adds an attribute with the same name to the serializer which overrides (inadvertently) core's definition of the attribute and results in issues such as users not being able to see the Review link if the plugin is disabled. This commit renames `can_review` to `can_review_code` everywhere in the plugin to avoid conflicts with core.