DEV: Use a single query to fetch the latest commit (#153)
Use a single GraphQL query to fetch the latest commit instead of two API calls as before. This also makes sure the job is not retried in case it failed the first time because that can flood Sidekiq with jobs as they are scheduled once per hour and repository.

This partially reverts commit 3de7cbbfbcb409dda9aaaab38c4dc6a4db54acb0.