FIX: Stop trying to sync repositories that have moved
We currently iterate over known GithubRepoCategory records and schedule
synchronization for each.

When a repo is renamed, the old GithubRepoCategory remains, albeit with
a repo_id that is null, alongside a new one. This causes a problem, as
when we try to synchronise it, the repo isn't there.

When a repo is transferred, the old GithubRepoCategory also remains,
causing the same problem.

This change fixes both cases by skipping repos that have been moved
when scheduling the synchronisation.