FIX: Don't associate pull requests from other repositories (#167)
As part of synchronising repositories we look at associated pull requests
of the commits in that repository. An assumption has been made here that
those pull requests will belong to the same repository that we are
synchronising.

This does not always hold true. In particular, when working with a forked
repo and integrating upstream changes, the commits brought in may have
associated pull requests from other repositories, causing synchronisation
to fail.

This change correctly assigns the owner and name of pull requests that
are associated with commits, and excludes pull requests from external
repositories by default.