PERF: Change sidekiq jobs to low queue (#199)
Why this change?

The jobs in this plugin can run for a long time depending on the amount
of changes that has happened in the repository since the last sync. If
there are a lot of changes, this job can end up taking a long time and
blocking other more important work on the `default` sidekiq queue.

Since the job is not critical to the function of the forum, we can put
the job in the `low` queue to depriortise it.